Split generic behavior impl type-arg tests
