+++ src/typechecker/tests/generic_behaviors.rs
@@ -4,5 +4,6 @@ mod collection;
 mod extends;
 mod generic_bounds;
 mod generic_function_diagnostics;
+mod impl_type_args;
 mod impls_and_requires;
 mod requires;
